Fix various compiler warnings (#815)

+++ Sources/CSFBAudioEngine/Decoders/SFBShortenDecoder.mm
@@ -1515,7 +1515,8 @@ NSError *genericShortenInvalidFormatErrorForURL(NSURL *_Nonnull url) noexcept {
         case functionBlocksize: {
             uint32_t uint = 0;
             if (!_input.getUInt32(uint, _version, static_cast<int>(std::log2(_blocksize))) || uint == 0 ||
-                uint > maxBlocksize || uint <= _wrap || static_cast<int>(uint) > _blocksize) {
+                uint > maxBlocksize || uint <= static_cast<uint32_t>(_wrap) ||
+                uint > static_cast<uint32_t>(_blocksize)) {
                 os_log_error(gSFBAudioDecoderLog, "Invalid or unsupported block size: %u", uint);
                 if (error) {
                     *error = [NSError errorWithDomain:SFBAudioDecoderErrorDomain
